requirements.txt
numpy==1.26.4
spacy==3.7.2
streamlit==1.39.0
pypdf==4.3.1
pdf2image==1.17.0
pillow==10.4.0
pytesseract==0.3.13
python-docx==1.1.2
reportlab==4.2.5
pyyaml==6.0.2
negspacy==1.0.2
